+++ migrations/20231217104821-create-cashback.js
@@ -20,7 +20,7 @@ module.exports = {
         onDelete: 'CASCADE',
       },
       amount: {
-        type: Sequelize.DOUBLE(20, 2),
+        type: Sequelize.DOUBLE,
         allowNull: false,
         defaultValue: 0
       },
