+++ migrations/20240122200555-create-user-referral.js
@@ -26,7 +26,7 @@ module.exports = {
         },
       },
       amount: {
-        type: Sequelize.DOUBLE(20, 2),
+        type: Sequelize.DOUBLE,
         allowNull: false,
         defaultValue: 0
       },
