+++ models/usercashback.js
@@ -20,7 +20,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     amount: {
-      type: DataTypes.DOUBLE(20, 2),
+      type: DataTypes.DOUBLE,
       allowNull: false,
       defaultValue: 0
     }
