+++ models/userreferral.js
@@ -25,7 +25,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     amount: {
-      type: DataTypes.DOUBLE(20, 2),
+      type: DataTypes.DOUBLE,
       defaultValue: 0
     }
   }, {
